WIP: cumulative cleanup changes from parallel agents
Includes assessments + source changes from agents 3, 4, 6, 7, 8
(unused code, circular deps, error handling, legacy code, comment cleanup).
Agents 1, 2, 5 pending. Final verify to run after all agents complete.
