+++ apps/loader/package.json
@@ -49,7 +49,6 @@
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
     "@vitejs/plugin-react": "^4.2.0",
-    "concurrently": "^8.2.0",
     "electron": "^28.0.0",
     "electron-builder": "^24.9.0",
     "eslint": "^8.57.1",
@@ -61,8 +60,7 @@
     "typescript": "^5.3.0",
     "vite": "^5.0.0",
     "vite-plugin-electron": "^0.15.0",
-    "vitest": "^1.0.0",
-    "wait-on": "^7.2.0"
+    "vitest": "^1.0.0"
   },
   "build": {
     "appId": "org.debrief.loader",
