+++ apps/web-shell/package.json
@@ -22,10 +22,8 @@
     "@debrief/utils": "workspace:*",
     "golden-layout": "^2.6.0",
     "leaflet": "^1.9.4",
-    "memfs": "^4.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "shared-zustand": "^2.1.0",
     "vscrui": "^0.1.0"
   },
   "devDependencies": {
