+++ apps/web-shell/src/tools/region/analysis/areaSummary.ts
@@ -4,7 +4,7 @@
  * Returns a bounding-box polygon with area statistics.
  */
 
-import type { MCPToolDefinition } from '../../../services/toolService';
+import type { MCPToolDefinition } from '@debrief/utils';
 import type { DebriefFeature } from '@debrief/schemas';
 import { OutputKindEnum } from '@debrief/schemas';
 import type { GeoJSONFeature } from '@debrief/utils';
