+++ apps/web-shell/src/tools/track/analysis/rangeBearing.ts
@@ -5,10 +5,9 @@
  * so the calcService can auto-open it in the Results panel.
  */
 
-import type { MCPToolDefinition } from '../../../services/toolService';
 import type { TrackFeature } from '@debrief/schemas';
 import { OutputKindEnum } from '@debrief/schemas';
-import type { GeoJSONFeature } from '@debrief/utils';
+import type { GeoJSONFeature, MCPToolDefinition } from '@debrief/utils';
 
 const EARTH_RADIUS_NM = 3440.065;
