+++ apps/web-shell/src/tools/track/analysis/trackStats.ts
@@ -4,10 +4,9 @@
  * total distance, and average speed.
  */
 
-import type { MCPToolDefinition } from '../../../services/toolService';
 import type { TrackFeature } from '@debrief/schemas';
 import { OutputKindEnum } from '@debrief/schemas';
-import type { GeoJSONFeature } from '@debrief/utils';
+import type { GeoJSONFeature, MCPToolDefinition } from '@debrief/utils';
 
 const EARTH_RADIUS_NM = 3440.065;
