+++ services/cli/debrief_cli/catalog.py
@@ -12,7 +12,7 @@ from pathlib import Path
 
 import click
 
-from debrief_cli.main import Context, pass_context
+from debrief_cli.context import Context, pass_context
 
 
 def _get_config_path() -> Path:
