+++ services/cli/debrief_cli/validate.py
@@ -11,7 +11,7 @@ import sys
 
 import click
 
-from debrief_cli.main import Context, pass_context
+from debrief_cli.context import Context, pass_context
 
 
 @click.command()
